Split project in start-client and start-site
The start.spring.io application currently relies on a default Web UI
built in. As of spring-io/initializr#854, the Initializr library does
not ship anymore with those defaults and developers need to implement
their own Controller and provide static assets to do so.

This commit splits the current start.spring.io in two modules:

* a start-site module that provides the service instance
* a start-client module that packages the client application into a JAR

For now, the Web UI is still being rendered partially on the server with
a templating engine. In future developments, the whole Web UI might be
rendered on the client side and the start-site module might only provide
JSON APIs to feed the client.