Restructure extension package
This commit restructures the package so that assets can be more easily
grouped and identified.

Customizers that are related to a dependency are located in an
sub-package of `dependency` with the name of the project. Code-related
customizers are located in the `code` package and Build-related
customizers in the `build` package.

There are other customizations of the `ProjectDescription` that should
not be registered in the `ProjectGenerationContext` located in the
`project` package. `META-INF/spring.factories` has been updated to
refer to the various elements rather than having a top-level
configuration that knows about everything.

Note that this is work in progress still as some refactoring of the
test infrastructure is in order to better isolate test assertions.

Closes gh-93