Copy bom repository metadata from spring cloud function
The spring cloud function customizer replaces the spring cloud function
with a dedicated artifact that doesn't have metadata. Because we remove
the original dependency, all its metadata is lost and won't be applied.

Previously, if such customizer applied, it didn't copy the repositories
associated with the bom of the spring cloud dependency if any. As a
result, if a repository was required it wasn't available anymore. This
commit makes sure to copy the bom repositories as well.

Closes gh-324