Add junit-jupiter testcontainers support
This commit adds the junit-jupiter module of testcontainers rather than
the "core" module as Junit Jupiter support in testcontainers is optional
and projects are using Junit Jupiter out-of-the-box.

Given that the module is always useful, we no longer remove it when a
supported entry is found, which reduces significantly the complexity of
the build customizer.

Closes gh-525