Fix Spring Cloud Contract Gradle plugin configuration
Previously, the generated configuration for the Spring Cloud Contract
Gradle plugin did not work with the latest version of the plugin.

- `targetFramework` was deprecated attribute and it is now deleted. As
the new attribute (`testFramework`) was available since at least Spring
Boot 2.2.x use it instead so the configuration can work with all version
after Spring Boot 2.2.x
- Since Spring Cloud Contract Gradle plugin 2.2.0.M1 the generated tests
are executed in separate task, `contractTest`. This tasks also need to
be configured to use JUnit Platform (JUnit 5). Otherwise the generated
tests are not going to be run.

See gh-613