Handle Spring Web and Spring Cloud Gateway properly
This commit adds a special check when Spring Cloud Gateway is selected
with Spring Web. The former does not support the latter, only Spring
WebFlux is supported at this time.

When such a combination is selected, the project is updated to use
WebFlux and a warning note is added in HELP.md.

Closes gh-653