Add Spring Cloud Azure MySQL/PostgreSQL starters
This commit adds new Azure starters, under certain conditions.

If the "azure-support" and "mysql" dependencies are selected, then the
"spring-cloud-azure-starter-jdbc-mysql" starter is added automatically
as well as a link to the relevant docs in the HELP file.

If the "azure-support" and "postgresql" dependencies are selected, then the
"spring-cloud-azure-starter-jdbc-postgresql" starter is added
automatically as well as a link to the relevant docs in the HELP file.

Closes gh-1056