Upgrade Spring Cloud Stream test dependency
Prior to this commit, the "spring-cloud-stream" dependency with the
"test-binder" classifier would be added to Spring Cloud Stream
applications for Maven builds only.

As of Spring Cloud Stream 4.0 (and Spring Boot 3.0), the coordinates of
this dependency changed to "spring-cloud-stream-test-binder" (without
any classifier). This commit ensures that the new coordinates are added
to Maven and Gradle builds for Spring Boot 3+ applications.

Closes gh-1069