Revert "Merge pull request #1113 from ZheSun88"
This reverts commit bae7f94cb3b4434f4a849e0bc5f31b209c29993a, reversing
changes made to 881a328664566cdc9e9acb9399241ae387d30b9c.

Unfortunately, when this was merged we did not realise that Hilla did
not support both Maven and Gradle. We do not have the necessary
infrastructure in place for a dependency to require a particular
build system. Until that infrastructure is in place or Hilla supports
both Maven and Gradle we don't want it to be an option on
start.spring.io.