Swap DockerImage for DockerService
This commit shares more metadata about a docker service as both the
docker compose and the testcontainers support need it. This introduces
a strategy interface to resolve a service based on an identifier. Future
version of the interface may include the chosen platform version if we
want to tune the image based on a Spring Boot generation.

Closes gh-1214