Raise minimum version to Spring Boot 2.6
This commit upgrades the service to reject requests to generate a
project with a version inferior to 2.6.0. As a result, any specific
customizations for those now unsupported versions have been removed.

This includes the swap between V1 and V2 version format as we no longer
need to accept the former.

See gh-1241