Add Netflix DGS Codegen plugin support
This commit adds a new dependency to start.spring.io, the "Netflix Data
Graph Service Codegen". This will not add a dependency to generated
project, but rather configure a Maven or a Gradle plugin in the build.

This Codegen plugin parses GraphQL schema files describing remote
GraphQL APIs and generates Java classes to easily create queries and
deserialize GraphQL responses.