Speed up the ProjectGenerationIntegrationTests
This also limits the maximum parallelism of that test to 4, as executing
this test on a machine with 32 cores will start 32 mavens / gradles all
downloading a fresh copy of the repository.

The Maven / Gradle homes are now created on demand and reused.
Additionally, they have predictable names in the temp folder and can be
reused for consecutive test runs.