Don't duplicate Postgres databases
PgVector is an extension for Postgres. Before this commit, we would add
a pgvector and a postgres Docker Compose service. The resulting app
doesn't start up, as there are now two JdbcConnectionDetails in the
context.

After that commit, we only add the pgvector Docker Compose service if
both Postgres and PgVector is selected.

Closes gh-1458