Validate and clean up all compatibility ranges
There are now no more unbounded ranges in the config, and Spring Cloud
has been set to be compatible with 3.2.x only.

The configuration is now consistently named 'compatibilityRange', and
the range doesn't include any spaces.

This removes some useless compatibilityRange declarations.

It also moves the compatibilityRange for Spring AI to an upper level.