Import testcontainers configuration in integration tests
This changes the code generation of the Testcontainers support: the
testcontainer beans are now defined in their own configuration. This
configuration is used in the test main method, and is @Import-ed in
the integration test.
The result is that the test main method starts testcontainers, and the
tests pass too as they now also start the testcontainers.

Closes gh-1247