Automatically add Spring Modulith event externalization dependency
If a message broker *and* Spring Modulith are declared as project
dependencies, the event externalization of Spring Modulith for
configured broker are now declared automatically as well. Also, the
Spring Modulith Events API dependency is declared in compile scope to
allow event types to be marked for externalization.

See gh-1528