Cache MetadataVerificationTests
They are quite slow to execute as they always resolve against an empty
Maven repository.

This commit moves the Homes and some other test support infra structure
to a shared module which is used in start-site and
start-site-verification in test scope.

The DependencyResolver used in MetadataVerificationTests is using the
Homes class, reusing the already cached repository, thus speeding it up.