Generate test app when RabbitMQ Streams is selected
Currently, when rabbitmq streams and docker compose/testcontainers
are selected, no container is provided. For that reason, the test and
test app can not be executed sucessfully. RabbitMQ Streams requires
an additional configuration to enable it using Spring Boot and also
in the container itself but providing RabbitMQ service would be enough
to execute the test and test app.