Add dependencies list to dependencies dialog
Users might misclic or change their mind about adding a dependency.
Having to close the dialog in order to remove the dependency is unnecessary
extra work.

This makes it so that the selected dependencies list is shown inside the
dialog in a non-obstrusive way so that users can see what they've chosen
so far and potentially remove them.